refactor: move SendVerificationTokenUseCase to verification token use cases folder
