+++ src/main/factories/user/external-sign-in/external-sign-in-factory.ts
@@ -7,7 +7,7 @@ import { JwtAdapter } from '@/infra/criptography';
 import { makeExternalSignInValidationFactory } from '@/main/factories/user';
 import { AddVerificationTokenUseCase } from '@/use-cases/verification-token/add-verification-token';
 import { EmailService } from '@/infra/services';
-import { SendVerificationTokenUseCase } from '@/use-cases/email-service/send-verification-token';
+import { SendVerificationTokenUseCase } from '@/use-cases/verification-token/send-verification-token';
 import env from '@/main/config/env';
 
 export function makeExternalSignInController(): IController {
