--- src/use-cases/email-service/errors/index.ts
@@ -1,1 +1,0 @@
-export * from '@/use-cases/email-service/errors/email-service-error';
