--- src/use-cases/email-service/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from '@/use-cases/email-service/interfaces/email-options-interface';
-export * from '@/use-cases/email-service/interfaces/email-service-interface';
