--- src/use-cases/email-service/send-verification-token/index.ts
@@ -1,1 +1,0 @@
-export * from '@/use-cases/email-service/send-verification-token/send-verification-token-use-case';
