+++ src/use-cases/errors/email-service-error.ts
