+++ src/use-cases/interfaces/email-options-interface.ts
