+++ src/use-cases/interfaces/email-service-interface.ts
