+++ src/use-cases/interfaces/index.ts
@@ -1,2 +1,4 @@
 export * from '@/use-cases/interfaces/use-case-interface';
 export * from '@/use-cases/interfaces/universally-unique-identifier-generator-interface';
+export * from '@/use-cases/interfaces/email-options-interface';
+export * from '@/use-cases/interfaces/email-service-interface';
