+++ src/use-cases/verification-token/send-verification-token/index.ts
@@ -1,0 +1,1 @@
+export * from '@/use-cases/verification-token/send-verification-token/send-verification-token-use-case';
